+++ Frends.AzureEventHub.Send/Frends.AzureEventHub.Send/Definitions/Result.cs
@@ -9,17 +9,17 @@ public class Result
     /// Indicates whether the messages were sent successfully.
     /// </summary>
     /// <example>true</example>
-    public bool Success { get; private set; }
+    public bool Success { get; init; }
 
     /// <summary>
     /// Message indicating the result of the operation.
     /// </summary>
     /// <example>A batch of 5 events has been published.</example>
-    public string Message { get; private set; }
+    public string Message { get; init; } = string.Empty;
 
-    internal Result(bool success, string message)
-    {
-        Success = success;
-        Message = message;
-    }
+    /// <summary>
+    /// Error information when Success is false and ThrowErrorOnFailure is set to false.
+    /// </summary>
+    /// <example>null</example>
+    public Error Error { get; init; }
 }